DiD regression with imputation method

DiD regression with imputation method

didImputation(
  y0,
  data,
  cohort,
  nevertreated.value = Inf,
  unit = NULL,
  time = NULL,
  coef = NULL,
  w = NULL,
  OATT = FALSE,
  se = T,
  tol = 1e-06,
  maxit = 100,
  mutatedata = FALSE,
  verbose = 0,
  effective.sample = 30,
  ...
)

Arguments

y0
Formula. model for Y(0). This is the full model without leads and lags used to predict the counterfactual outcome.

data
A data frame or data.table

cohort
Character. Time of treatment identifier. By default it expect Inf for never treated units. You can override this with nevertreated.value

nevertreated.value
Any, default is Inf. Value encoding the cohort of never treated units.

unit
Character, default is NULL. Unit identifier. If NULL, it default to the first fixed effect.

time
Character, default is NULL. Time period identifier. If NULL, it default to the second fixed effect.

coef
Numeric vector of leads and lags (eg coef = c(-5, 8)). By default estimates all lags available.

w
(not yet implemented) Numerical or Variable name. Default is NULL. Estimation weights

OATT
Logical, default is FALSE. If TRUE then the overall average treatment effect is computed instead of ATT at each horizon.

se
Logical, default is TRUE. Should standard errors be reported

tol
Numeric, default is 1e-6. Tolerance level for weights convergence.

maxit
Numeric, default is 100. Maximum number of iterations for computing weights

mutatedata
Logical, default is FALSE. USE AT YOUR OWN RISK. This option will modify your data in place instead of making a copy of it which could be useful with large dataset if your RAM is limited. It may result in lost observations and bugs.

verbose
Numeric, default is 0. Level of verbosity.

effective.sample
Numeric, default is 30. Effective sample size under which the function will throw a warning. See details.

Value

A didImputation object with the results of the imputation estimation.

data

Data used for estimation.

convergence_time

User and CPU time for weights convergence.

pvalue

p-value for the positive horizon estimates.

coeftable

Table of regression coefficients.

wald

Wald statistic of the pre-trend regression.

coefs

Average treatment effects from the imputation procedure.

niterations

Number of iterations it took to compute the weights.

pre_trends

fixest regression object for the pre-trends estimation.

References

Borusyak, K., Jaravel, X., & Spiess, J. (2021). Revisiting event study designs: Robust and efficient estimation. Working paper.

Stata implementation from the authors:

did_imputation (https://github.com/borusyak/did_imputation)

See also

Author

Antoine Mayerowitz

Maxime Gravoueille